requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai-whisper==20231117
transformers==4.40.0
torch==2.3.0
sentencepiece==0.2.0
sacremoses==0.1.1
IndicTransToolkit
edge-tts==6.1.9
nest_asyncio==1.6.0
librosa==0.10.2
soundfile==0.12.1
noisereduce==3.0.2
numpy==1.26.4
langdetect==1.0.9
huggingface-hub==0.23.0
accelerate==0.30.0